refactor: User 엔티티를 UserDetails 에서 분리하고 보안 필터에서 DTO 사용
