+++ src/main/java/com/github/shCHO9801/climbing_record_app/user/entity/User.java
@@ -11,8 +11,6 @@ import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -20,8 +18,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 @Entity
 @Table(name = "users")
@@ -32,7 +28,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 @Builder
 @ToString(exclude = "password")
 @EqualsAndHashCode(of = {"userNum", "id", "email"})
-public class User implements UserDetails {
+public class User {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -65,41 +61,6 @@ public class User implements UserDetails {
 
   @Column(nullable = false)
   private LocalDateTime updatedAt;
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    return List.of(() -> "ROLE_" + role.name());
-  }
-
-  @Override
-  public String getUsername() {
-    return this.id;
-  }
-
-  @Override
-  public String getPassword() {
-    return this.password;
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return UserDetails.super.isAccountNonExpired();
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return UserDetails.super.isAccountNonLocked();
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return UserDetails.super.isCredentialsNonExpired();
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return UserDetails.super.isEnabled();
-  }
 
   @PrePersist
   protected void onCreate() {
