+++ src/main/java/com/github/shCHO9801/climbing_record_app/user/service/CustomUserDetailsService.java
@@ -3,11 +3,10 @@ package com.github.shCHO9801.climbing_record_app.user.service;
 import static com.github.shCHO9801.climbing_record_app.exception.ErrorCode.USER_NOT_FOUND;
 
 import com.github.shCHO9801.climbing_record_app.exception.CustomException;
+import com.github.shCHO9801.climbing_record_app.user.auth.dto.CustomerUserDetails;
 import com.github.shCHO9801.climbing_record_app.user.entity.User;
 import com.github.shCHO9801.climbing_record_app.user.repository.UserRepository;
-import java.util.Collections;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -22,15 +21,9 @@ public class CustomUserDetailsService implements UserDetailsService {
   @Override
   public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
     User user = userRepository.findByUsername(username)
-        .orElseThrow(() -> new CustomException(USER_NOT_FOUND)); // 추후 커스텀 익셉션 추가 예정
+        .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
 
-    return new org.springframework.security.core.userdetails.User(
-        user.getUsername(),
-        user.getPassword(),
-        Collections.singletonList(
-            new SimpleGrantedAuthority("ROLE_" + user.getRole().name())
-        )
-    );
+    return CustomerUserDetails.fromEntity(user);
   }
 
 }
